Language Cleanup and Suggestions
The majority of the intro is fine but feels “limiting”, and I’m not
sure adding more words will help.

For example:
* There is lots of talk about “acquisition” and design/deployment - but
it’s missing language around ongoing management and operations.
* Also, there is talk about IT’s “HW, Firmware, OS, Middleware, and
SW…” but the language overlaps the policies, practices, and operational
behaviors (i.e. soft-side of IT).

Overall, I think this (and the document as a whole) would benefit from
“plain language” review and in particular a focus to remove the “legacy
vernacular” (major/minor, IT, DME/SS, …) that is implicit in a rewrite.