Non IT Specific
Too much “IT” specific language - the CIO (or delegate) should be able
to provide context on “how it should work” (e.g. maybe a procurement /
payment cycle) independently of the “IT specific portions” (e.g. this
limits the CIO to just saying “I don’t think it should be JAVA”).